+basic map loading

+++ C10Engine/cPlayer.h
@@ -17,9 +17,9 @@
 
 
 // Player physics constants
-const float PLAYER_ACCELERATION = 0.6;
-const float PLAYER_MAX_SPEED = 4;
-const float PLAYER_FRICTION = 0.4;
+const float PLAYER_ACCELERATION = (float)0.6;
+const float PLAYER_MAX_SPEED = (float)4;
+const float PLAYER_FRICTION = (float)0.4;
